optional-requirements.txt
auto-sklearn==0.3.0
pynisher

# We need lasagne 0.2dev1 + corresponding theano version
theano==0.8.2
git+https://github.com/Lasagne/Lasagne.git@master

torch
torchvision
xgboost
